requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
Werkzeug==3.0.1
